% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\docType{data}
\name{ptclr_n}
\alias{ptclr_n}
\alias{ptclr_d}
\alias{ptclr_l}
\alias{Colours}
\title{Colours}
\format{
An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
ptclr_n

ptclr_d

ptclr_l
}
\description{
Normal theme blue in hex

Dark theme blue in hex

Light theme blue in hex
}
\concept{Theme colours}
\keyword{datasets}
