% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{contrast}
\alias{contrast}
\title{Contrast Colour Selector}
\usage{
contrast(colour, dark = "black", light = "white")
}
\arguments{
\item{colour}{A character vector of colour names or hex codes.}

\item{dark}{A character string specifying the colour to use for dark contrast (default is "black").}

\item{light}{A character string specifying the colour to use for light contrast (default is "white").}
}
\value{
A character vector of the same length as `colour`, containing the contrast colours (either `dark` or `light`).
}
\description{
This function determines the appropriate contrast colour (either dark or light) for a given set of colours.
}
\examples{
contrast(c("grey20", "grey80"))
contrast(c("#000000", "#FFFFFF"), dark = "navy", light = "yellow")
}
