% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gh_link}
\alias{gh_link}
\title{Produce Github link in Markdown}
\usage{
gh_link(
  text = NULL,
  path,
  repo,
  user = "petrbouchal",
  remote = NULL,
  branch = "main"
)
}
\arguments{
\item{text}{Link text}

\item{path}{path to file (the bit that goes after /tree/{branch},
so presumably the local path from project root)}

\item{repo}{repo name; used if `remote` is not provided.}

\item{user}{Github user; used if `remote` is not provided. Defaults to `"petrbouchal"`}

\item{remote}{remote in the form of user/repo.}

\item{branch}{branch. Default `"main"`}
}
\value{
character vector of same length as inputs

Inputs are recycled if not all of the same length
}
\description{
Useful for linking to files on Github that live in a local clone of a repo
}
