% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhub_to_drat.R
\name{insert_package_into_drat}
\alias{insert_package_into_drat}
\title{Put source and binary package into drat}
\usage{
insert_package_into_drat(
  pkg,
  last_tag = T,
  repodir = getOption("dratRepo", default = "~/github/drat"),
  build_win = T
)
}
\arguments{
\item{pkg}{path to package directory}

\item{last_tag}{whether to check out the latest tag (which should point to a release)}

\item{repodir}{directory containing drat repo. Looks for the `dratRepo` option, if not found, defaults to "~/github/drat"}

\item{build_win}{(TRUE) whether to build the Windows binary on rhub and put it into drat.}
}
\value{
TRUE on success
}
\description{
Build source and Mac binary bundle locally and Windows binary on rhub and put them into a drat repository.
}
\examples{
\dontrun{
insert_package_to_drat("~/github/czso", last_tag = T, build_win = T)
}
# ADD_EXAMPLES_HERE
}
