% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandoc-quotes.R
\name{pandoc_quotes}
\alias{pandoc_quotes}
\title{Insert Pandoc argument into YAML to use quotes filter}
\usage{
pandoc_quotes()
}
\value{
RETURN_DESCRIPTION
}
\description{
To be used as the `pandoc_args` YAML param in Rmd documents. Configure by setting the `lang` YAML param ("en"/"cs") in top-level YAML
}
\details{
See [LUA filter manual](https://github.com/pandoc/lua-filters/blob/master/pandoc-quotes.lua/man/pandoc-quotes.lua.md) for more granular settings.
}
\examples{
# output:
#   html_document:
#     pandoc_args: !expr ptrr::pandoc_quotes()
}
