% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown-output.R
\name{ptrr_word}
\alias{ptrr_word}
\title{Basic Schola Empirica word document}
\usage{
ptrr_word(...)
}
\arguments{
\item{...}{Arguments to be passed to `[bookdown::word_document2]`}
}
\value{
A modified `word_document2` with the standard Schola formatting.
}
\description{
This is a function called in the output of the yaml of the Rmd file to
specify using the standard Schola word document formatting.
}
\examples{
\dontrun{
  output: ptrr::ptrr_word
}
}
\concept{Report templates and formats}
