% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts-ragg.R
\name{register_plex_mono}
\alias{register_plex_mono}
\title{Register IBM font with system}
\usage{
register_plex_mono(family = "IBM Plex Mono", features = NULL, warn = TRUE)
}
\arguments{
\item{family}{What you want the resulting font family to be called in the registry. Defaults to "IBM Plex Sans Condensed"}

\item{features}{Unused for now}

\item{warn}{Whether to warn when a font family of the same name already exists. Defaults to TRUE.}
}
\value{
TRUE if a font was successfully registered, FALSE otherwise. Called for side effects.
}
\description{
Register the IBM Plex Sans Condensed font files contained in the package with the OS font system. Wrapper around `systemfonts::register_font()`.
Useful in situations where you cannot install fonts into the system font location, e.g. on CI systems.
}
\examples{
register_plexsans_condensed("ibmps")
}
