% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{set_geom_defaults}
\alias{set_geom_defaults}
\title{Set geom defaults for ptrr package/theme}
\usage{
set_geom_defaults(color = "blue")
}
\arguments{
\item{color}{color, defaults to "blue"}
}
\description{
Set geom defaults for ptrr package/theme
}
