% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-components.R
\name{use_ptrr_site}
\alias{use_ptrr_site}
\title{FUNCTION_TITLE}
\usage{
use_ptrr_site(path = ".", ghactions = TRUE)
}
\arguments{
\item{path}{DESCRIPTION.}

\item{ghactions}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
