% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-components.R
\name{use_ptrr_slides}
\alias{use_ptrr_slides}
\title{FUNCTION_TITLE}
\usage{
use_ptrr_slides(path = "docs")
}
\arguments{
\item{path}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
