% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{vis_palette}
\alias{vis_palette}
\title{Visualize Color Palettes}
\usage{
vis_palette(palette_list)
}
\arguments{
\item{palette_list}{A (named) list of color palettes. Each element of the list should be a vector of color codes, optionally with names.}
}
\value{
A `ggplot` object visualizing the color palettes.
}
\description{
This function takes a list of color palettes and visualizes them using `ggplot2`.
}
\details{
The colour palettes inside the list can be named, as can be the individual colours.
The names are then displayed in the swatch plot.
}
\examples{
palettes <- list(
  palette1 = c(moc = "#FF0000", trochu = "#00FF00", malo = "#0000FF"),
  palette2 = c("#FFFF00", "#FF00FF", "#00FFFF")
)
vis_palette(palettes)
}
