% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown-output.R
\name{ptrr_word}
\alias{ptrr_word}
\title{Basic ptrr word document}
\usage{
ptrr_word(
  reference_docx = find_resource("ptrr_word", "template.docx"),
  number_sections = FALSE,
  ...
)
}
\arguments{
\item{...}{Arguments to be passed to `[bookdown::word_document2]`}
}
\value{
A modified `word_document2` with the standard ptrr formatting.
}
\description{
This is a function called in the output of the yaml of the Rmd file to
specify using the standard ptrr word document formatting. By default,
it uses the Word template built into the package as the `reference_docx`
YAML option
}
\examples{
\dontrun{
  output: ptrr::ptrr_word
}
}
\concept{Report templates and formats}
