% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{render_readme}
\alias{render_readme}
\title{FUNCTION_TITLE}
\usage{
render_readme(
  path = "README.Rmd",
  output_format = "github_document",
  output_file = "README.md"
)
}
\arguments{
\item{path}{DESCRIPTION.}

\item{output_format}{DESCRIPTION.}

\item{output_file}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
