% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-components.R
\name{use_ptrr_docs}
\alias{use_ptrr_docs}
\title{FUNCTION_TITLE}
\usage{
use_ptrr_docs(file = "docs", template = "docs", package = "ptrr", edit = FALSE)
}
\arguments{
\item{file}{DESCRIPTION.}

\item{template}{DESCRIPTION.}

\item{package}{DESCRIPTION.}

\item{edit}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
